Move FunctionControlFlowInfo from opt to gen
CFG construction is a property of the IR, not an optimization pass.
Moving it to gen allows any pass (translation, codegen, liveness) to
use CFG data without importing opt.
