+++ gen/function_control_flow_info.go
@@ -1,21 +1,20 @@
-package opt
+package gen
 
 import (
 	"alon.kr/x/graph"
-	"alon.kr/x/usm/gen"
 )
 
 // FunctionControlFlowInfo holds the flattened basic-block list, the
 // block-to-index mapping, and the control-flow graph for a function. It is
-// derived purely from gen.FunctionInfo and can be used by any pass that needs
+// derived purely from FunctionInfo and can be used by any pass that needs
 // CFG traversal (e.g. constant propagation, liveness analysis) or block-index
 // lookups (e.g. SSA construction).
 type FunctionControlFlowInfo struct {
 	// A linear representation of all basic blocks in the function.
-	BasicBlocks []*gen.BasicBlockInfo
+	BasicBlocks []*BasicBlockInfo
 
 	// Maps each basic block to its index in BasicBlocks.
-	BasicBlocksToIndex map[*gen.BasicBlockInfo]uint
+	BasicBlocksToIndex map[*BasicBlockInfo]uint
 
 	ControlFlowGraph *graph.Graph
 }
@@ -23,10 +22,10 @@ type FunctionControlFlowInfo struct {
 // NewFunctionControlFlowInfo builds a FunctionControlFlowInfo from the given
 // function by linearising its basic blocks and constructing the CFG from their
 // forward edges.
-func NewFunctionControlFlowInfo(function *gen.FunctionInfo) FunctionControlFlowInfo {
+func NewFunctionControlFlowInfo(function *FunctionInfo) FunctionControlFlowInfo {
 	basicBlocks := function.CollectBasicBlocks()
 
-	blockToIndex := make(map[*gen.BasicBlockInfo]uint, len(basicBlocks))
+	blockToIndex := make(map[*BasicBlockInfo]uint, len(basicBlocks))
 	for i, b := range basicBlocks {
 		blockToIndex[b] = uint(i)
 	}
